[HTML] Improvements to attribute completions (#2433)
* [HTML] Allow the user to type > at the end of a tag to "overwrite"

If for some reason a user wants a second >, they can press that key
twice.

* [HTML] Improve attribute completion flow

Always adds a field to attribute completions to provide a better typing
experience for users. This allows the user to press [tab] instead of
typing a close quote, or close quote and greater than symbol.